refactor: authentication.rs to impl of Bunkr
