+++ programs/vault-program/src/instructions/mod.rs
@@ -5,7 +5,6 @@ pub mod thaw_non_fungible;
 pub mod lock_fungible;
 pub mod unlock_fungible;
 pub mod close_bunkr;
-pub mod authentication;
 pub mod test_withdraw;
 pub mod change_withdrawal_address;
 pub mod reset_root;
@@ -18,7 +17,6 @@ pub use thaw_non_fungible::*;
 pub use lock_fungible::*;
 pub use unlock_fungible::*;
 pub use close_bunkr::*;
-pub use authentication::*;
 pub use test_withdraw::*;
 pub use change_withdrawal_address::*;
 pub use reset_root::*;
