+++ programs/vault-program/src/instructions/test_withdraw.rs
@@ -3,14 +3,6 @@ use {
     anchor_lang::{prelude::*}
 };
 
-
-
-
-pub fn handler(ctx: Context<TestWithdraw>) -> Result<()> {
-    msg!("Withdrawal successful");
-    Ok(())
-}
-
 #[derive(Accounts)]
 pub struct TestWithdraw<'info> {
     #[account(mut, seeds=[b"bunkr", signer.key().as_ref()], bump)]
